fix: Sửa lỗi đăng nhập - Cập nhật BCrypt password encoder
BREAKING CHANGES:
- Đổi PasswordEncoder từ plain text sang BCryptPasswordEncoder
- Cập nhật mật khẩu trong database sang BCrypt hash

Backend Changes:
- SecurityConfig.java: Thay đổi PasswordEncoder sang BCryptPasswordEncoder
- application.properties: Sửa SQL Server port từ 65356 → 1433
- PasswordHashGenerator.java: Thêm utility để generate BCrypt hash

Frontend Changes:
- api-config.js: Sửa API URL từ port 8080 → 8082

Database Changes:
- Nhom8_DoAnPhanMem.sql: Thêm script update mật khẩu sang BCrypt

Tài khoản demo (password không đổi):
- admin@gmail.com / admin123
- anhthu@gmail.com / tp123
- nv01@crm.vn / nv01123

Fixes: #issue-login-authentication

+++ BE/src/main/java/com/nhom8/crm/controller/TrungLapController.java
@@ -32,7 +32,7 @@ public class TrungLapController {
                 .map(pair -> {
                     KhachHang k1 = (KhachHang) pair[0];
                     KhachHang k2 = (KhachHang) pair[1];
-                    int similarity = khachHangService.calculateSimilarity(k1.getHoTen(), k2.getHoTen());
+                    double similarity = khachHangService.calculateSimilarity(k1.getHoTen(), k2.getHoTen());
 
                     return TrungLapKhachHangResponse.builder()
                             .id(k1.getMaKhachHang() + "_" + k2.getMaKhachHang())
