+++ BE/src/main/java/com/nhom8/crm/dto/response/TrungLapKhachHangResponse.java
@@ -10,7 +10,7 @@ public class TrungLapKhachHangResponse {
     private String id;
     private CustomerDetail customer1;
     private CustomerDetail customer2;
-    private Integer similarity;
+    private Double similarity;
 
     @Data
     @Builder
